Properly load translations
Apparently the correct way to load translations is to add the
Text Domain and Domain Path header and call load_plugin_textdomain
when the plugin_loaded action hook is executed [1]. It is unclear
whether it would be better to call load_plugin_textdomain when the init
action hook is executed, e.g. Jetpack uses some kind of hack to use
both.

[1] http://ottopress.com/2013/language-packs-101-prepwork/